marshalling system for lists has been implemented!
ok so i had to make some tweeks to get this to work nicely:
-time is now stored as seconds since epoch rather than as datetime objects
-a lot of optional arguments in list and card def have been eliminated, instead those values are initialized as  whatever is desired as the default value and then during loading we create the card/list and then edit those attributes after

why do we have to do it this way? by putting the default value in the arguments section scoping says all instances are actually taking the SAME list as an argument which leads to collisions and leaking that mess everything up. try it for yourself

save and load are now methods of List:
to save a list merely call list.save and it saves to the lists folder in storage
to load a saved list:
-create an empty list
-using that list call list.load(saved_list_title)
--list.load then finds the file labeled with the right title and loads the data into the new list